feat: renamed and removed files, testing more queries
