--- api/src/__tests__/mutations.test.js
@@ -1,5 +1,0 @@
-describe("Placeholder test", () => {
-    test("1 + 1", () => {
-        expect(1 + 1).toBe(2);
-    });
-});
